+++ apps/macos/metro.config.js
@@ -3,27 +3,29 @@
  * integration tests during local development or on CI services.
  */
 
-const path = require('path');
-const blacklist = require('metro-config/src/defaults/blacklist');
-const { getWatchFolders } = require('@uifabricshared/build-native');
+const { defaultWatchFolders, exclusionList } = require('@rnx-kit/metro-config');
 const { getDefaultConfig } = require('metro-config');
+const path = require('path');
 
-const rnmPath = path.dirname(require.resolve('react-native-macos/package.json'));
+const blockList = exclusionList([
+  // Exclude other test apps
+  /.*\/apps\/(?:android|ios|web|win32|windows)\/.*/,
+  // Exclude build output directory
+  /.*\/apps\/macos\/dist\/.*/,
+]);
 
 module.exports = (async () => {
   const {
     resolver: { sourceExts, assetExts },
   } = await getDefaultConfig();
   return {
-    watchFolders: getWatchFolders(),
+    watchFolders: defaultWatchFolders(__dirname),
     resolver: {
       assetExts: assetExts.filter((ext) => ext !== 'svg'),
       sourceExts: [...sourceExts, 'svg'],
-      extraNodeModules: {
-        'react-native': rnmPath,
-      },
       platforms: ['macos', 'ios'],
-      blacklistRE: blacklist([/node_modules\/react-native\/.*/]),
+      blacklistRE: blockList,
+      blockList,
     },
     transformer: {
       babelTransformerPath: require.resolve('react-native-svg-transformer'),