+++ packages/theming/theme-tokens/src/light/tokens-alias.ts
@@ -1,0 +1,3 @@
+import aliasTokens from './reactnative/tokens-aliases.json';
+
+export default aliasTokens;