+++ packages/components/Link/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/link
 
-This log was last generated on Wed, 21 Jul 2021 22:55:40 GMT and should not be manually modified.
+This log was last generated on Mon, 26 Jul 2021 20:41:04 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.12.19
+
+Mon, 26 Jul 2021 20:41:04 GMT
+
+### Patches
+
+- add new use-slot package, move framework to consume that package (jasonmo@microsoft.com)
+
+## 0.12.18
+
+Sun, 25 Jul 2021 16:30:17 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/adapters to v0.6.5 (email not defined)
 
 ## 0.12.14