+++ packages/components/MenuButton/CHANGELOG.md
@@ -1,8 +1,25 @@
 # Change Log - @fluentui-react-native/menu-button
 
-This log was last generated on Fri, 23 Jul 2021 08:20:11 GMT and should not be manually modified.
+This log was last generated on Mon, 26 Jul 2021 20:41:04 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.2.14
+
+Mon, 26 Jul 2021 20:41:04 GMT
+
+### Patches
+
+- add new use-slot package, move framework to consume that package (jasonmo@microsoft.com)
+- Small change in menu button snapshot name (krsiler@microsoft.com)
+
+## 0.2.13
+
+Sun, 25 Jul 2021 16:30:17 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/button to v0.17.15 (email not defined)
 
 ## 0.2.10