+++ packages/theming/theme-tokens/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/theme-tokens
 
-This log was last generated on Wed, 21 Jul 2021 22:55:40 GMT and should not be manually modified.
+This log was last generated on Mon, 26 Jul 2021 20:41:04 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.3.1
+
+Mon, 26 Jul 2021 20:41:04 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/theme-types to v0.9.1 (jasonmo@microsoft.com)
+
+## 0.3.0
+
+Sat, 24 Jul 2021 00:28:25 GMT
+
+### Minor changes
+
+- Add more aslias tokens, regenerate files (email not defined)
 
 ## 0.2.4