+++ packages/theming/theme-types/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/theme-types
 
-This log was last generated on Fri, 23 Jul 2021 17:28:27 GMT and should not be manually modified.
+This log was last generated on Mon, 26 Jul 2021 20:41:04 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.9.1
+
+Mon, 26 Jul 2021 20:41:04 GMT
+
+### Patches
+
+- add new use-slot package, move framework to consume that package (jasonmo@microsoft.com)
+
+## 0.9.0
+
+Sat, 24 Jul 2021 00:28:25 GMT
+
+### Minor changes
+
+- createAliasTokens to generate alias token structure (email not defined)
 
 ## 0.8.3