+++ packages/utils/adapters/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/adapters
 
-This log was last generated on Thu, 22 Jul 2021 20:55:49 GMT and should not be manually modified.
+This log was last generated on Mon, 26 Jul 2021 20:41:04 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.6.6
+
+Mon, 26 Jul 2021 20:41:04 GMT
+
+### Patches
+
+- add new use-slot package, move framework to consume that package (jasonmo@microsoft.com)
+
+## 0.6.5
+
+Sun, 25 Jul 2021 16:30:17 GMT
+
+### Patches
+
+- Fix build error (email not defined)
 
 ## 0.6.4